+++ Shoot/src/CacheFile.cpp
@@ -110,7 +110,7 @@ namespace shoot
 				break;
 			}
 		}
-		return dest.length();
+		return (s32)dest.length();
 	}
 
 	//! reads a line until a delimiter
